feat: challenges 레이아웃 및 컴포넌트 변경 중
