+++ src/app/(detail)/stats/api/stats/route.ts
