+++ src/app/(detail)/stats/page.tsx
